+++ backend/src/resource/asset/index.html
@@ -1,15 +1,8 @@
 <!DOCTYPE html><html lang="en"><head>
         <meta charset="utf-8">
         <title>70年代、80年代、90年代，怀旧、复古、回忆</title>
-        <link rel="stylesheet" href="/asset/purecss.min-b565cdb90b7aac9f.css">
+        <link rel="stylesheet" href="/asset/bootstrap.min-6c05f7396999b9df.css">
         <link rel="stylesheet" href="/asset/tag-4d2c4d8d06ebde6.css">
-        <style type="text/css">
-            body{padding:20px}
-            .tar{text-align:right}
-            .tac{text-align:center}
-            .tag-btn{color:white;border-radius:4px;text-shadow:0 1px 1px rgba(0, 0, 0, 0.2);background:rgb(28, 184, 65);margin-right:6px}
-            .tag-link{margin-right:6px}
-        </style>
-    <script type="module">import init from '/asset/index-d1a376e8d6448cde.js';init('/asset/index-d1a376e8d6448cde_bg.wasm');</script></head>
+    <script type="module">import init from '/asset/index-1fa1e498e77bbea2.js';init('/asset/index-1fa1e498e77bbea2_bg.wasm');</script></head>
     <body>
 </body></html>
