+++ backend/src/resource/page/config.html
@@ -3,44 +3,44 @@
 <head>
     <meta charset="UTF-8">
     <title>配置</title>
-    <link rel="stylesheet" href="/asset/purecss.min-b565cdb90b7aac9f.css">
+    <link rel="stylesheet" href="/asset/bootstrap.min-6c05f7396999b9df.css">
 </head>
 <body>
-<div class="pure-g">
-    <div class="pure-u-1-1">
-        <h1>博客信息配置</h1>
+<div class="container">
+    <div class="row">
+        <div class="col">
+            <h1>博客信息配置</h1>
+        </div>
     </div>
-</div>
-<div class="pure-g">
-    <div class="pure-u-1-3"></div>
-    <div class="pure-u-1-3">
-        <div id="errorMessage" style="color:red;display:none"></div>
-        <form class="pure-form pure-form-aligned">
-            <fieldset>
-                <div class="pure-control-group">
+    <div class="row">
+        <div class="col-md"></div>
+        <div class="col-md">
+            <div id="errorMessage" style="color:red;display:none"></div>
+            <form class="row g-3">
+                <div class="col-12">
                     <label for="aligned-domain">博客根域名</label>
-                    <input type="text" id="aligned-domain" placeholder="博客根域名" />
+                    <input type="text" id="aligned-domain" class="form-control" placeholder="博客根域名"/>
                 </div>
-                <div class="pure-control-group">
+                <div class="col-12">
                     <label for="aligned-name">博客名称</label>
-                    <input type="text" id="aligned-name" placeholder="博客名称" />
+                    <input type="text" id="aligned-name" class="form-control" placeholder="博客名称"/>
                 </div>
-                <div class="pure-control-group">
+                <div class="col-12">
                     <label for="aligned-copyright">版权信息</label>
-                    <input type="password" id="aligned-copyright" placeholder="版权信息" />
+                    <input type="password" id="aligned-copyright" class="form-control" placeholder="版权信息"/>
                 </div>
-                <div class="pure-control-group">
+                <div class="col-12">
                     <label for="aligned-license">备案信息</label>
-                    <input type="email" id="aligned-license" placeholder="备案信息" />
+                    <input type="email" id="aligned-license" class="form-control" placeholder="备案信息"/>
                 </div>
-                <div class="pure-controls">
-                    <button type="submit" class="pure-button pure-button-primary">更新</button>
-                    <button type="button" class="pure-button" onclick="location.href='/';">返回到博客</button>
+                <div class="col-12">
+                    <button type="submit" class="btn btn-primary">更新</button>
+                    <button type="button" class="btn" onclick="location.href='/';">返回到博客</button>
                 </div>
-            </fieldset>
-        </form>
+            </form>
+        </div>
+        <div class="col-md"></div>
     </div>
-    <div class="pure-u-1-3"></div>
 </div>
 </body>
 </html>
