+++ backend/src/resource/page/login.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <title>配置管理用户</title>
-    <link rel="stylesheet" href="/asset/purecss.min-b565cdb90b7aac9f.css">
+    <link rel="stylesheet" href="/asset/bootstrap.min-6c05f7396999b9df.css">
     <script type="text/javascript">
         function login() {
             const user = {
@@ -37,41 +37,38 @@
     </script>
 </head>
 <body>
-<div class="pure-g">
-    <div class="pure-u-1-1">
-        <h1>管理登录</h1>
+<div class="container">
+    <div class="row">
+        <div class="col">
+            <h1>管理登录</h1>
+        </div>
     </div>
-</div>
-<div class="pure-g">
-    <div class="pure-u-1-3"></div>
-    <div class="pure-u-1-3">
-        <div id="errorMessage" style="color:red;display:none"></div>
-        <form class="pure-form pure-form-aligned" onsubmit="login();return false;">
-            <fieldset>
-                <div class="pure-control-group">
-                    <label for="aligned-email">邮箱地址</label>
-                    <input type="email" id="aligned-email" placeholder="邮箱地址"/>
+    <div class="row">
+        <div class="col-md"></div>
+        <div class="col-md">
+            <div id="errorMessage" style="color:red;display:none"></div>
+            <form class="row g-3" onsubmit="login();return false;">
+                <div class="col-12">
+                    <label for="aligned-email" class="form-label">邮箱地址</label>
+                    <input type="email" id="aligned-email" class="form-control" placeholder="邮箱地址"/>
                 </div>
-                <div class="pure-control-group">
-                    <label for="aligned-password">登录密码</label>
-                    <input type="password" id="aligned-password" placeholder="密码"/>
+                <div class="col-12">
+                    <label for="aligned-password" class="form-label">登录密码</label>
+                    <input type="password" id="aligned-password" class="form-control" placeholder="密码"/>
                 </div>
-                <div class="pure-control-group">
-                    <label for="aligned-captcha">验证码</label>
-                    <input type="text" id="aligned-captcha" placeholder="验证码"/>
+                <div class="col-md-4">
+                    <label for="aligned-captcha" class="form-label">验证码</label>
+                    <input type="text" id="aligned-captcha" class="form-control" placeholder="验证码"/>
+                    <img src="/tool/verify-image"/>
                 </div>
-                <div class="pure-control-group">
-                    <label></label>
-                    <img src="/tool/verify-image" />
+                <div class="col-12">
+                    <button type="submit" class="btn btn-primary">登录</button>
+                    <button type="button" class="btn" onclick="location.href='/';">返回到博客</button>
                 </div>
-                <div class="pure-controls">
-                    <button type="submit" class="pure-button pure-button-primary">登录</button>
-                    <button type="button" class="pure-button" onclick="location.href='/';">返回到博客</button>
-                </div>
-            </fieldset>
-        </form>
+            </form>
+        </div>
+        <div class="col-md"></div>
     </div>
-    <div class="pure-u-1-3"></div>
 </div>
 </body>
 </html>
