+++ backend/src/resource/page/reg.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <title>配置管理用户</title>
-    <link rel="stylesheet" href="/asset/purecss.min-b565cdb90b7aac9f.css">
+    <link rel="stylesheet" href="/asset/bootstrap.min-6c05f7396999b9df.css">
     <script type="text/javascript">
         function register() {
             const user = {
@@ -27,41 +27,43 @@
     </script>
 </head>
 <body>
-<div class="pure-g">
-    <div class="pure-u-1-1">
-        <h1>创建管理员</h1>
+<div class="container">
+    <div class="row">
+        <div class="col">
+            <h1>创建管理员</h1>
+        </div>
     </div>
-</div>
-<div class="pure-g">
-    <div class="pure-u-1-3"></div>
-    <div class="pure-u-1-3">
-        <div id="errorMessage" style="color:red;display:none"></div>
-        <form class="pure-form pure-form-aligned" onsubmit="register();return false;">
-            <fieldset>
-                <div class="pure-control-group">
-                    <label for="aligned-email">邮箱地址</label>
-                    <input type="email" id="aligned-email" placeholder="邮箱地址" />
-                </div>
-                <div class="pure-control-group">
-                    <label></label>
-                    <span class="pure-form-message-inline">用于管理博客或忘记密码之用，如果忘记管理员账号，可以删除<</span>
-                </div>
-                <div class="pure-control-group">
-                    <label for="aligned-password1">登录密码</label>
-                    <input type="password" id="aligned-password1" placeholder="密码" />
-                </div>
-                <div class="pure-control-group">
-                    <label for="aligned-password2">确认密码</label>
-                    <input type="password" id="aligned-password2" placeholder="密码" />
-                </div>
-                <div class="pure-controls">
-                    <button type="submit" class="pure-button pure-button-primary">创建管理员</button>
-                    <button type="button" class="pure-button" onclick="location.href='/';">返回到博客</button>
-                </div>
-            </fieldset>
-        </form>
+    <div class="row">
+        <div class="col-md"></div>
+        <div class="col-md">
+            <div id="errorMessage" style="color:red;display:none"></div>
+            <form class="row g-3" onsubmit="register();return false;">
+                <fieldset>
+                    <div class="col-12">
+                        <label for="aligned-email" class="form-label">邮箱地址</label>
+                        <input type="email" id="aligned-email" class="form-control" placeholder="邮箱地址"/>
+                    </div>
+                    <div class="col-12">
+                        <label></label>
+                        <span class="pure-form-message-inline">用于管理博客或忘记密码之用，如果忘记管理员账号，可以删除<</span>
+                    </div>
+                    <div class="col-12">
+                        <label for="aligned-password1" class="form-label">登录密码</label>
+                        <input type="password" id="aligned-password1" class="form-control" placeholder="密码"/>
+                    </div>
+                    <div class="col-12">
+                        <label for="aligned-password2" class="form-label">确认密码</label>
+                        <input type="password" id="aligned-password2" class="form-control" placeholder="密码"/>
+                    </div>
+                    <div class="col-12">
+                        <button type="submit" class="btn btn-primary">创建管理员</button>
+                        <button type="button" class="btn" onclick="location.href='/';">返回到博客</button>
+                    </div>
+                </fieldset>
+            </form>
+        </div>
+        <div class="col-md"></div>
     </div>
-    <div class="pure-u-1-3"></div>
 </div>
 </body>
 </html>
