+++ backend/src/service/asset_list.rs
@@ -1,9 +1,9 @@
 [
+("bootstrap.min-6c05f7396999b9df.css", include_bytes!(r#"..\resource\asset/bootstrap.min-6c05f7396999b9df.css.gz"#), "text/css"),
 ("codemirror.min.css", include_bytes!(r#"..\resource\asset/codemirror.min.css.gz"#), "text/css"),
-("index-d1a376e8d6448cde.js", include_bytes!(r#"..\resource\asset/index-d1a376e8d6448cde.js.gz"#), "text/javascript"),
-("index-d1a376e8d6448cde_bg.wasm", include_bytes!(r#"..\resource\asset/index-d1a376e8d6448cde_bg.wasm.gz"#), "application/wasm"),
+("index-1fa1e498e77bbea2.js", include_bytes!(r#"..\resource\asset/index-1fa1e498e77bbea2.js.gz"#), "text/javascript"),
+("index-1fa1e498e77bbea2_bg.wasm", include_bytes!(r#"..\resource\asset/index-1fa1e498e77bbea2_bg.wasm.gz"#), "application/wasm"),
 ("index.html", include_bytes!(r#"..\resource\asset/index.html.gz"#), "text/html; charset=utf-8"),
-("purecss.min-b565cdb90b7aac9f.css", include_bytes!(r#"..\resource\asset/purecss.min-b565cdb90b7aac9f.css.gz"#), "text/css"),
 ("snippets/blog-frontend-3607fd663da6040e/asset/editor.js", include_bytes!(r#"..\resource\asset/snippets\blog-frontend-3607fd663da6040e\asset\editor.js.gz"#), "text/javascript"),
 ("tag-4d2c4d8d06ebde6.css", include_bytes!(r#"..\resource\asset/tag-4d2c4d8d06ebde6.css.gz"#), "text/css"),
 ("toastui-editor-all.min.js", include_bytes!(r#"..\resource\asset/toastui-editor-all.min.js.gz"#), "text/javascript"),
