+++ frontend/index.html
@@ -3,18 +3,11 @@
     <head>
         <meta charset="utf-8">
         <title>70年代、80年代、90年代，怀旧、复古、回忆</title>
-        <link data-trunk rel="css" href="asset/purecss.min.css">
+        <link data-trunk rel="css" href="asset/bootstrap.min.css">
         <link data-trunk rel="css" href="asset/tag.css">
         <link data-trunk rel="copy-file" href="asset/codemirror.min.css">
         <link data-trunk rel="copy-file" href="asset/toastui-editor.min.css">
         <link data-trunk rel="copy-file" href="asset/toastui-editor-all.min.js">
-        <style type="text/css">
-            body{padding:20px}
-            .tar{text-align:right}
-            .tac{text-align:center}
-            .tag-btn{color:white;border-radius:4px;text-shadow:0 1px 1px rgba(0, 0, 0, 0.2);background:rgb(28, 184, 65);margin-right:6px}
-            .tag-link{margin-right:6px}
-        </style>
     </head>
     <body></body>
 </html>
